staging: ncipher: More initial comment related changes.
Specifically the following items were addressed:
  osif.h has been streamlined and mostly eliminated, with a new solo.h used instead.
  A mutex is now used instead of a spinlock across nfp_ioctl
  Resolved previous PR2 comments regarding c99 .style= assignments and
nfp_copy_to/from_user methods eliminated.
  And changed more variables into explicitly sized data types
