+++ drivers/staging/ncipher/osif.c
@@ -5,10 +5,10 @@
  *
  */
 
-#include <linux/spinlock.h>
-#include <linux/slab.h>
+#include "solo.h"
 
-#include "osif.h"
+/* error conversion ------------------------------------------- */
+int     nfp_error(int oerr);
 
 void nfp_sleep(int ms)
 {
@@ -22,31 +22,7 @@ void nfp_sleep(int ms)
 	finish_wait(&q, &wait);
 }
 
-int nfp_config_inl(struct nfp_dev *ndev, int offset, u32 *res)
-{
-	if (!ndev || !ndev->pcidev)
-		return NFP_ENODEV;
-	pci_read_config_dword(ndev->pcidev, offset, res);
-	return 0;
-}
-
 /* user space memory access ---------------------------------- */
-
-int nfp_copy_from_user(char *kbuf, const char *ubuf, int len)
-{
-	int oserr;
-
-	oserr = copy_from_user(kbuf, ubuf, len) ? -EFAULT : 0;
-	return nfp_error(oserr);
-}
-
-int nfp_copy_to_user(char *ubuf, const char *kbuf, int len)
-{
-	int oserr;
-
-	oserr = copy_to_user(ubuf, kbuf, len) ? -EFAULT : 0;
-	return nfp_error(oserr);
-}
 
 int nfp_copy_from_user_to_dev(struct nfp_dev *ndev, int bar, int offset,
 			      const char *ubuf, int len)
@@ -78,35 +54,6 @@ int nfp_copy_to_dev(struct nfp_dev *ndev, int bar,
 {
 	memcpy(ndev->bar[bar] + offset, kbuf, len);
 	return NFP_SUCCESS;
-}
-
-/* pci fixed length accessors. The below functions are used predominantly
- * to access CSR registers in pci memory space.
- */
-u32 nfp_inl(struct nfp_dev *ndev, int bar, int offset)
-{
-	nfp_log(NFP_DBG3, "%s: addr %p", __func__, ndev->bar[bar] + offset);
-	return ioread32(ndev->bar[bar] + offset);
-}
-
-u16 nfp_inw(struct nfp_dev *ndev, int bar, int offset)
-{
-	nfp_log(NFP_DBG3, "%s: addr %p", __func__, ndev->bar[bar] + offset);
-	return ioread16(ndev->bar[bar] + offset);
-}
-
-void nfp_outl(struct nfp_dev *ndev, int bar, int offset, u32 data)
-{
-	nfp_log(NFP_DBG3, "%s: addr %p, data %x", __func__,
-		ndev->bar[bar] + offset, data);
-	iowrite32(data, ndev->bar[bar] + offset);
-}
-
-void nfp_outw(struct nfp_dev *ndev, int bar, int offset, u16 data)
-{
-	nfp_log(NFP_DBG3, "%s: addr %p, data %x", __func__,
-		ndev->bar[bar] + offset, data);
-	iowrite16(data, ndev->bar[bar] + offset);
 }
 
 /* logging ---------------------------------------------------- */
@@ -141,12 +88,7 @@ void nfp_log(int level, const char *fmt, ...)
 	}
 }
 
-struct errstr {
-	int oserr;
-	int nferr;
-};
-
-static struct errstr errtab[] = { { -EFAULT, NFP_EFAULT },
+struct errstr errtab[] = { { -EFAULT, NFP_EFAULT },
 				  { -ENOMEM, NFP_ENOMEM },
 				  { -EINVAL, NFP_EINVAL },
 				  { -EIO, NFP_EIO },
@@ -167,19 +109,4 @@ int nfp_error(int oserr)
 		perr++;
 	}
 	return NFP_EUNKNOWN;
-}
-
-int nfp_oserr(int nferr)
-{
-	struct errstr *perr;
-
-	if (nferr == NFP_SUCCESS)
-		return 0;
-	perr = errtab;
-	while (perr->nferr) {
-		if (perr->nferr == nferr)
-			return perr->oserr;
-		perr++;
-	}
-	return -EIO;
 }
