+++ drivers/staging/ncipher/solo.h
@@ -1,7 +1,8 @@
 /* SPDX-License-Identifier: GPL-2.0+ */
 /** @file
  *
- * Declares the nFast PCI register interface.
+ * Declares the nFast PCI register interface, driver runtime
+ * structures, and supporting items.
  *
  * The interface presented by nFast PCI devices consists of:
  *
@@ -12,8 +13,27 @@
  *
  */
 
-#ifndef NFP_PCI_H
-#define NFP_PCI_H
+#ifndef SOLO_H
+#define SOLO_H
+
+#include <linux/version.h>
+#include <linux/types.h>
+#include <linux/time.h>
+#include <linux/io.h>
+#include <linux/sched.h>
+#include <linux/kernel.h>
+#include <linux/string.h>
+#include <linux/timer.h>
+#include <linux/ptrace.h>
+#include <linux/errno.h>
+#include <linux/ioport.h>
+#include <linux/slab.h>
+#include <linux/interrupt.h>
+#include <linux/pci.h>
+#include <linux/poll.h>
+#include <linux/mutex.h>
+#include <linux/spinlock.h>
+#include <linux/nshield_solo.h>
 
 /* Sizes of some regions */
 
@@ -189,4 +209,145 @@
 #define NFPCI_SCRATCH_ERROR_LO		    2
 #define NFPCI_SCRATCH_ERROR_HI		    3
 
-#endif /* NFP_PCI_H */
+#define NFP_BARSIZES_COUNT 6
+#define NFP_BARSIZES_MASK ~0xF
+
+/* per-instance device structure ------------------------------------ */
+struct nfp_dev {
+	/* downward facing part of the device interface */
+	u8 *bar[NFP_BARSIZES_COUNT];
+	void *extra[NFP_BARSIZES_COUNT];
+	int busno;
+	int slotno;
+	void *cmdctx;
+	char *iobuf;
+	int active_bar;
+	int created;
+	int conn_status;
+	int detection_type;
+	int iosize[6];
+	u32 irq;
+	struct nfpcmd_dev const *cmddev;
+	struct nfdev_stats_str stats;
+
+	/* upward facing part of the device interface */
+	u8 *read_buf;
+	dma_addr_t read_dma;
+
+	u8 *write_buf;
+	dma_addr_t write_dma;
+
+	struct pci_dev *pcidev;
+
+	int busy;
+	int ifvers;
+	struct timer_list rd_timer;
+
+	wait_queue_head_t rd_queue;
+	unsigned long rd_ready;
+	unsigned long rd_outstanding;
+	int rd_ok;
+
+	wait_queue_head_t wr_queue;
+	unsigned long wr_ready;
+	unsigned long wr_outstanding;
+	int wr_ok;
+
+	spinlock_t spinlock;     /* protect this struct */
+	struct mutex ioctl_mutex;  /* lock across ioctl */
+};
+
+/* Per-device-type command handlers --------------------------------- */
+struct nfpcmd_dev {
+	const char *name;
+	u16 vendorid, deviceid, sub_vendorid, sub_deviceid;
+	u32 bar_sizes[NFP_BARSIZES_COUNT]; /* includes IO bit */
+	u32 flags, max_ifvers;
+
+	int (*create)(struct nfp_dev *ndev);
+	int (*destroy)(void *ctx);
+	int (*started)(struct nfp_dev *ndev, int lock_flag);
+	int (*stopped)(struct nfp_dev *ndev);
+	int (*open)(void *ctx);
+	int (*close)(void *ctx);
+	int (*isr)(void *ctx, int *handled);
+	int (*write_block)(u32 addr, const char *ublock, int len, void *ctx);
+	int (*read_block)(char *ublock, int len, void *ctx, int *rcount);
+	int (*channel_update)(char *data, int len, void *ctx);
+	int (*ensure_reading)(u32 addr, int len, void *ctx, int lock_flag);
+	int (*debug)(int cmd, void *ctx);
+	int (*setcontrol)(const struct nfdev_control_str *control,
+			  void *ctx); /* may be NULL */
+	int (*getstatus)(struct nfdev_status_str *status,
+			 void *ctx); /* may be NULL */
+};
+
+/* These instances are defined in the per-board driver modules */
+extern const struct nfpcmd_dev i21555_cmddev;
+extern const struct nfpcmd_dev fsl_t1022_cmddev;
+
+/* user and device memory space access ---------------------------- */
+/*
+ * NB these 2 functions are not guaranteed to be re-entrant for a given device
+ */
+int nfp_copy_from_user_to_dev(struct nfp_dev *ndev, int bar, int offset,
+			      const char *ubuf, int len);
+int nfp_copy_to_user_from_dev(struct nfp_dev *ndev, int bar, int offset,
+			      char *ubuf, int len);
+
+int nfp_copy_from_dev(struct nfp_dev *ndev, int bar, int offset,
+		      char *kbuf, int len);
+int nfp_copy_to_dev(struct nfp_dev *ndev, int bar, int offset,
+		    const char *kbuf, int len);
+
+#define NFP_CMD_FLG_NEED_MSI   0x2
+
+#ifndef PCI_BASE_ADDRESS_SPACE_PREFETCHABLE
+#define PCI_BASE_ADDRESS_SPACE_PREFETCHABLE 0x8
+#endif
+
+/*
+ * This masks off the bottom bits of the PCI_CSR_BAR which signify that the
+ * BAR is an IO BAR rather than a MEM BAR
+ */
+#define NFP_MEMBAR_MASK ~0xf
+
+/* callbacks from command drivers -------------------------------------- */
+void nfp_read_complete(struct nfp_dev *ndev, int ok);
+void nfp_write_complete(struct nfp_dev *ndev, int ok);
+
+/* debug ------------------------------------------------------------ */
+
+#define NFP_DBG1	1
+#define NFP_DBGE	NFP_DBG1
+#define NFP_DBG2	2
+#define NFP_DBG3	3
+#define NFP_DBG4	4
+
+void nfp_log(int severity, const char *format, ...);
+extern int nfp_debug;
+
+/* internal error codes --------------------------------------------- */
+#define NFP_SUCCESS	0x0
+#define NFP_EFAULT	0x1
+#define NFP_ENOMEM	0x2
+#define NFP_EINVAL	0x3
+#define NFP_EIO		0x4
+#define NFP_ENXIO	0x5
+#define NFP_ENODEV	0x6
+#define NFP_EINTR	0x7
+#define NFP_ESTARTING	0x8
+#define NFP_EAGAIN	0x9
+#define NFP_EPOLLING	0xA
+#define NFP_EINTERRUPT	0xB
+#define NFP_EUNKNOWN	0x100
+
+/* error conversions table */
+struct errstr {
+	int oserr;
+	int nferr;
+};
+
+extern struct errstr errtab[];
+
+#endif /* SOLO_H */
